[MC-26] [web] refactor: move folder mobile to numerical-guidance
