+++ apps/web/app/store/querys/numerical-guidance/indicator-quote.query.ts
